+++ openeogeotrellis/deploy/submit_batch_job_spark3.sh
@@ -101,7 +101,8 @@ sparkDriverJavaOptions="-Dscala.concurrent.context.maxThreads=2 -Dpixels.treshol
  -Dlog4j2.configurationFile=file:/opt/venv/openeo-geopyspark-driver/batch_job_log4j2.xml\
  -Dhdp.version=3.1.4.0-315\
  -Dsoftware.amazon.awssdk.http.service.impl=software.amazon.awssdk.http.urlconnection.UrlConnectionSdkHttpService\
- -Dopeneo.logging.threshold=$logging_threshold"
+ -Dopeneo.logging.threshold=$logging_threshold\
+ --add-opens=java.base/sun.net.util=ALL-UNNAMED"
 
 sparkExecutorJavaOptions="-Dlog4j2.configurationFile=file:/opt/venv/openeo-geopyspark-driver/batch_job_log4j2.xml\
  -Dsoftware.amazon.awssdk.http.service.impl=software.amazon.awssdk.http.urlconnection.UrlConnectionSdkHttpService\
@@ -225,6 +226,9 @@ spark-submit \
  --conf spark.extraListeners=org.openeo.sparklisteners.LogErrorSparkListener,org.openeo.sparklisteners.BatchJobProgressListener \
  --conf spark.sql.adaptive.coalescePartitions.parallelismFirst=false \
  --conf spark.sql.adaptive.advisoryPartitionSizeInBytes=5242880 \
+ --conf spark.yarn.jars=local:///usr/local/spark/jars/* \
+ --conf spark.driver.userClassPathFirst=false \
+ --conf spark.executor.userClassPathFirst=false \
  --files "${files}" \
  --py-files "${pyfiles}" \
  --conf spark.hadoop.security.authentication=kerberos --conf spark.yarn.maxAppAttempts=1 \