+++ tests/deploy/test_batch_job.py
@@ -19,7 +19,7 @@ from openeo_driver.dry_run import DryRunDataTracer
 from openeo_driver.save_result import ImageCollectionResult
 from openeo_driver.testing import DictSubSet, ListSubSet
 from openeo_driver.utils import read_json
-from osgeo import gdal
+import osgeo.gdal
 from pytest import approx
 from shapely.geometry import box, mapping, shape
 
@@ -853,8 +853,8 @@ def reproject_raster_file(source_path: str, destination_path: str, dest_crs: str
         Height of the output raster in pixels.
         Same remark as for `width`: need to specify output raster's width and height.
     """
-    opts = gdal.WarpOptions(dstSRS=dest_crs, width=width, height=height)
-    gdal.Warp(destNameOrDestDS=destination_path, srcDSOrSrcDSTab=source_path, options=opts)
+    opts = osgeo.gdal.WarpOptions(dstSRS=dest_crs, width=width, height=height)
+    osgeo.gdal.Warp(destNameOrDestDS=destination_path, srcDSOrSrcDSTab=source_path, options=opts)
 
 
 @mock.patch("openeo_driver.ProcessGraphDeserializer.evaluate")